Update p.c
This is the first stage of implementation after creating the base code for Player vs bot. Most bot methods are implemented, and only basic checking and handling is required. Moving forward, we'll code the game such that the difficulties in an ability is done in the same method and not in 3 different methods( easy, medium, hard).